project 3 fix

+++ project3/src/main/java/ru/alishev/springcourse/project3/models/Sensor.java
@@ -1,9 +1,10 @@
 package ru.alishev.springcourse.project3.models;
 
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 
-import javax.persistence.*;
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.Size;
 import java.util.List;
 
 @Entity
@@ -12,7 +13,7 @@ public class Sensor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "name")
     @NotEmpty(message = "Название сенсора не должно быть пустым")
@@ -28,11 +29,11 @@ public class Sensor {
 
     public Sensor() {}
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
